requirement.txt
django==2.2.19
django-filter==2.0.0
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
pillow==8.1.2